Replace (potential) leading dash with underscore
As noted elsewhere, I am not sure whether my old notes
indicated this change was needed to fix an actual problem
or guard against a hypothetical issue that could occur
when resetting the default password to the new one (e.g.,
shell interpolation of special chars).

Making the change for the Go implementation on the off-chance
that this is a real problem.

refs GH-15